Refactor SimulationProxy and SimulationHost.
Because one window can at maximum display one simulation,
a lot of code from simulation proxy can be removed.
Additionally all listeners in SimultaionProxy could be
replaced by direct method calls the UI class.

Additionally the SimualtionHost is no longer required,
because the lifetime of the application is now the same
as the liftetime of the Simulation.

Closes #137